data types

+++ com.revolsys.open.core/src/main/java/com/revolsys/util/number/Floats.java
@@ -82,7 +82,7 @@ public class Floats extends AbstractDataType {
   @Override
   protected Object toObjectDo(final Object value) {
     final String string = DataTypes.toString(value);
-    return Double.valueOf(string);
+    return Float.valueOf(string);
   }
 
   @Override
